+++ HISTORY.md
@@ -12,6 +12,7 @@
 * gui: fixed exporting media only for a single scene
 * gui: fixed cloud file format components
 * gui: "projectized" cloud file format for a single scene
+* gui: fixed cloud file format for multi-scene projects 
 
 ### 2021-05-11
 * gui: add multi-scene projects